move some test_block machinery to armi.testing * test_block.buildSimpleFuelBlock ->   testing.buildSimpleFuelHexBlock * test_block.buildLinkedFuelBlock ->   testing.buildLinkedFuelHexblock * test_block.loadTestBlock -> testing.buildComplexTestBlock * NUM_PINS_IN_TEST_BLOCK (test_blocks -> testing) * applyDummyData (test_blocks -> testing)
