+++ armi/reactor/tests/test_hexBlockRotate.py
@@ -18,11 +18,11 @@ import math
 import unittest
 
 import numpy as np
+from testing import NUM_PINS_IN_TEST_BLOCK, buildComplexHexBlock
 
 from armi.reactor.blocks import HexBlock
 from armi.reactor.components import Component
 from armi.reactor.grids import CoordinateLocation, HexGrid, IndexLocation, MultiIndexLocation
-from armi.reactor.tests.test_blocks import NUM_PINS_IN_TEST_BLOCK, loadTestBlock
 from armi.utils import iterables
 
 
@@ -48,7 +48,7 @@ class HexBlockRotateTests(unittest.TestCase):
     PIN_DATA = np.arange(NUM_PINS_IN_TEST_BLOCK, dtype=float)
 
     def setUp(self):
-        self.baseBlock = loadTestBlock()
+        self.baseBlock = buildComplexHexBlock()
         self._assignParamData(self.BOUNDARY_PARAMS, self.BOUNDARY_DATA)
         self._assignParamData(self.PIN_PARAMS, self.PIN_DATA)
